requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
feedparser==6.0.11
lxml==5.1.0
PyNaCl==1.5.0
